+++ prodotti_espresso/egg-info/history.rst
@@ -1,7 +1,14 @@
+12.0.1.0.6 (2023-05-30)
+~~~~~~~~~~~~~~~~~~~~~~~
+
+* [IMP] Flag espresso on sale order line form / Form righe ordini con flag prodotto espresso
+* [IMP] Flag espresso on sale order form / Form ordini con flag prodotto espresso
+* [IMP] Sale order BI with espresso filter / Filtro espresso in BI ordini
+
 12.0.1.0.5 (2023-02-08)
 ~~~~~~~~~~~~~~~~~~~~~~~
 
-* [IMP] flag espresso on invoice line / righe fattura con flag prodotto espresso
+* [IMP] Flag espresso on invoice line / righe fattura con flag prodotto espresso
 
 12.0.1.0.4 (2023-01-20)
 ~~~~~~~~~~~~~~~~~~~~~~~